+++ plugin/flags_sources/flags_file.py
@@ -4,7 +4,7 @@ Attributes:
     log (logging.Logger): current logger.
 """
 from .flags_source import FlagsSource
-from ..tools import File
+from ..utils.file import File
 from ..utils.singleton import FlagsFileCache
 from ..utils.flag import Flag
