+++ tests/test_c_cpp_properties.py
@@ -4,7 +4,7 @@ from os import path, environ
 from unittest import TestCase
 
 from EasyClangComplete.plugin.flags_sources import c_cpp_properties
-from EasyClangComplete.plugin import tools
+from EasyClangComplete.plugin.utils import tools
 from EasyClangComplete.plugin.utils import flag
 from EasyClangComplete.plugin.utils import search_scope
