+++ tests/test_compilation_db.py
@@ -4,7 +4,7 @@ from os import path
 from unittest import TestCase
 
 from EasyClangComplete.plugin.flags_sources import compilation_db
-from EasyClangComplete.plugin import tools
+from EasyClangComplete.plugin.utils import tools
 from EasyClangComplete.plugin.utils import flag
 from EasyClangComplete.plugin.utils import search_scope
 from EasyClangComplete.plugin.utils import singleton
