+++ tests/test_error_vis.py
@@ -8,8 +8,8 @@ from EasyClangComplete.plugin.error_vis import popup_error_vis
 from EasyClangComplete.plugin.settings import settings_manager
 from EasyClangComplete.plugin.settings import settings_storage
 from EasyClangComplete.plugin.popups import popups
-from EasyClangComplete.plugin import view_config
-from EasyClangComplete.plugin import tools
+from EasyClangComplete.plugin.view_config import view_config_manager
+from EasyClangComplete.plugin.utils import action_request
 
 from EasyClangComplete.tests import gui_test_wrapper
 
@@ -18,16 +18,16 @@ imp.reload(gui_test_wrapper)
 imp.reload(popup_error_vis)
 imp.reload(settings_manager)
 imp.reload(settings_storage)
-imp.reload(view_config)
+imp.reload(view_config_manager)
 imp.reload(popups)
-imp.reload(tools)
+imp.reload(action_request)
 
-ActionRequest = tools.ActionRequest
+ActionRequest = action_request.ActionRequest
 PopupErrorVis = popup_error_vis.PopupErrorVis
 GuiTestWrapper = gui_test_wrapper.GuiTestWrapper
 SettingsManager = settings_manager.SettingsManager
 SettingsStorage = settings_storage.SettingsStorage
-ViewConfigManager = view_config.ViewConfigManager
+ViewConfigManager = view_config_manager.ViewConfigManager
 Popup = popups.Popup
 
 test_file = namedtuple('test_file', 'name')
