+++ tests/test_view_config.py
@@ -5,8 +5,9 @@ from os import path
 
 from EasyClangComplete.plugin.settings import settings_manager
 from EasyClangComplete.plugin.settings import settings_storage
-from EasyClangComplete.plugin import view_config
-from EasyClangComplete.plugin import tools
+from EasyClangComplete.plugin.view_config import view_config
+from EasyClangComplete.plugin.view_config import view_config_manager
+from EasyClangComplete.plugin.utils import tools
 
 from EasyClangComplete.tests import gui_test_wrapper
 
@@ -14,6 +15,7 @@ imp.reload(gui_test_wrapper)
 imp.reload(settings_manager)
 imp.reload(settings_storage)
 imp.reload(view_config)
+imp.reload(view_config_manager)
 imp.reload(tools)
 
 SettingsManager = settings_manager.SettingsManager
